+++ components/merkledb/src/lib.rs
@@ -155,7 +155,7 @@ pub use self::{
 #[doc(no_inline)]
 pub use self::{
     proof_list_index::{ListProof, ProofListIndex},
-    proof_map_index::{MapProof, ProofMapIndex},
+    proof_map_index::{HashedProofMap, MapProof, ProofMapIndex, RawProofMap},
 };
 
 #[macro_use]