+++ examples/cryptocurrency-advanced/backend/src/schema.rs
@@ -14,12 +14,11 @@
 
 //! Cryptocurrency database schema.
 
-use exonum_merkledb::{IndexAccess, ObjectHash, ProofListIndex, ProofMapIndex};
+use exonum_merkledb::{IndexAccess, ObjectHash, ProofListIndex, RawProofMap};
 
 use exonum::crypto::{Hash, PublicKey};
 
 use crate::{wallet::Wallet, INITIAL_BALANCE};
-use exonum_merkledb::proof_map_index::RawProofMap;
 
 /// Database schema for the cryptocurrency.
 #[derive(Debug)]